requirements.txt
argparse
numpy
pyyaml
requests
tqdm
wandb